project/requirements.txt
cryptography==38.0.4
cryptography.egg==info
fastapi==0.110.0
psycopg2_binary==2.9.9
python-dotenv==1.0.1